+++ .travis.yml
@@ -9,7 +9,7 @@ env:
     global:
         - PHPUNIT_FLAGS="-v"
         - SYMFONY_PHPUNIT_DIR="$HOME/symfony-bridge/.phpunit"
-        - DEPENDENCIES="zendframework/zend-diactoros:^1.4.1"
+        - DEPENDENCIES="zendframework/zend-diactoros:^1.4.1 http-interop/http-factory-diactoros:^1.0"
 
 matrix:
     fast_finish: true
@@ -26,8 +26,11 @@ matrix:
           dist: 'precise'
           env: DEPENDENCIES=""
         - php: 5.4
+          env: DEPENDENCIES="zendframework/zend-diactoros:^1.4.1"
         - php: 5.5
+          env: DEPENDENCIES="zendframework/zend-diactoros:^1.4.1"
         - php: 5.6
+          env: DEPENDENCIES="zendframework/zend-diactoros:^1.4.1"
         - php: 7.0
         - php: 7.1
         - php: 7.2
@@ -36,9 +39,9 @@ matrix:
           # Test LTS versions. This makes sure we do not use Symfony packages with version greater
           # than 2 or 3 respectively.
         - php: 7.2
-          env: DEPENDENCIES="symfony/lts:^2 symfony/force-lowest:~2.8.0 zendframework/zend-diactoros:^1.4.1"
+          env: DEPENDENCIES="$DEPENDENCIES symfony/lts:^2 symfony/force-lowest:~2.8.0"
         - php: 7.2
-          env: DEPENDENCIES="symfony/lts:^3 symfony/force-lowest:~3.4.0 zendframework/zend-diactoros:^1.4.1"
+          env: DEPENDENCIES="$DEPENDENCIES symfony/lts:^3 symfony/force-lowest:~3.4.0"
 
           # Latest commit to master
         - php: 7.2
